back-end/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.3.1
django-ninja==1.3.0
pydantic==2.10.6
pydantic_core==2.27.2
sqlparse==0.5.3
typing_extensions==4.12.2
